+++ SplunkAgent/Sources/SplunkAgent/Public API/Modules/AppStart/Proxies/Module/AppStart+InternalTracking.swift
@@ -16,6 +16,7 @@ limitations under the License.
 */
 
 import Foundation
+@_spi(SplunkInternal) internal import SplunkAppStart
 
 extension AppStart {
 
@@ -23,11 +24,14 @@ extension AppStart {
 
     @discardableResult
     func track(didBecomeActive: Date, didFinishLaunching: Date?, willEnterForeground: Date?) -> any AppStartModule {
-        module.track(
-            didBecomeActive: didBecomeActive,
-            didFinishLaunching: didFinishLaunching,
-            willEnterForeground: willEnterForeground
-        )
+        module.track(didBecomeActive: didBecomeActive, didFinishLaunching: didFinishLaunching, willEnterForeground: willEnterForeground)
+
+        return self
+    }
+
+    @discardableResult
+    func track(initialLifecycle snapshot: AppStartLifecycleSnapshot) -> any AppStartModule {
+        module.track(initialLifecycle: snapshot)
 
         return self
     }